Clean-up representation and uflacs
