+++ doc/source/index.rst
@@ -31,11 +31,6 @@ API reference
    ffcx.optimization
    ffcx.representationutils
 
-.. autosummary::
-   :toctree: _autogenerated
-
-   ffcx.uflacs
-
 
 Indices and tables
 ==================
