+++ ffcx/codegeneration/codegeneration.py
@@ -32,29 +32,12 @@ code_blocks = namedtuple("code_blocks", ["elements", "dofmaps",
 def generate_code(ir, parameters):
     """Generate code blocks from intermediate representation."""
 
-    logger.debug("Compiler stage 4: Generating code")
-
     # Generate code for finite_elements
-    logger.debug("Generating code for {} finite_element(s)".format(len(ir.elements)))
     code_finite_elements = [finite_element_generator(element_ir, parameters) for element_ir in ir.elements]
-
-    # Generate code for dofmaps
-    logger.debug("Generating code for {} dofmap(s)".format(len(ir.dofmaps)))
     code_dofmaps = [dofmap_generator(dofmap_ir, parameters) for dofmap_ir in ir.dofmaps]
-
-    # Generate code for coordinate_mappings
-    logger.debug("Generating code for {} coordinate_mapping(s)".format(len(ir.coordinate_mappings)))
     code_coordinate_mappings = [coordinate_mapping_generator(cmap_ir, parameters) for cmap_ir in ir.coordinate_mappings]
-
-    # Generate code for integrals
-    logger.debug("Generating code for integrals")
     code_integrals = [integral_generator(integral_ir, parameters) for integral_ir in ir.integrals]
-
-    # Generate code for forms
-    logger.debug("Generating code for forms")
     code_forms = [form_generator(form_ir, parameters) for form_ir in ir.forms]
-
-    logger.debug("Generating code for expressions")
     code_expressions = [expression_generator(expression_ir, parameters) for expression_ir in ir.expressions]
 
     return code_blocks(elements=code_finite_elements, dofmaps=code_dofmaps,
