+++ ffcx/codegeneration/evalderivs.py
@@ -1,4 +1,4 @@
-"""Work in progress translation of FFC evaluatebasis code to uflacs CNodes format."""
+"""Work in progress translation of FFC evaluatebasis code to CNodes format."""
 
 import logging
 
@@ -334,7 +334,7 @@ def generate_tabulate_dmats(L, dofs_data):
             matrix[i, ...] = numpy.transpose(dmat)
 
         # TODO: Use precision from parameters here
-        from ffcx.ir.uflacs.elementtables import clamp_table_small_numbers
+        from ffcx.ir.elementtables import clamp_table_small_numbers
         matrix = clamp_table_small_numbers(matrix)
 
         # O(n^2) matrix matching...
