+++ ffcx/ir/representation.py
@@ -28,6 +28,11 @@ from ffcx.fiatinterface import (EnrichedElement, FlattenedDimensions,
                                 MixedElement, QuadratureElement, SpaceOfReals,
                                 create_element)
 from ffcx.ir import dof_permutations
+from ffcx.ir.representationutils import (QuadratureRule,
+                                         create_quadrature_points_and_weights)
+from ffcx.ir.integral import compute_integral_ir
+from ufl.sorting import sorted_expr_sum
+from ufl.classes import Integral
 from FIAT.hdiv_trace import HDivTrace
 
 logger = logging.getLogger(__name__)
@@ -69,7 +74,7 @@ ir_coordinate_map = namedtuple('ir_coordinate_map', ['id', 'prefix', 'name', 'si
                                                      'coordinate_finite_element_classname',
                                                      'scalar_coordinate_finite_element_classname',
                                                      'scalar_dofmap_name'])
-ir_integral = namedtuple('ir_integral', ['representation', 'integral_type', 'subdomain_id',
+ir_integral = namedtuple('ir_integral', ['integral_type', 'subdomain_id',
                                          'rank', 'geometric_dimension', 'topological_dimension',
                                          'entitytype', 'num_facets', 'num_vertices', 'needs_oriented',
                                          'enabled_coefficients', 'element_dimensions', 'element_ids',
@@ -96,8 +101,6 @@ def compute_ir(analysis: namedtuple, object_names, prefix, parameters, visualise
     """Compute intermediate representation.
 
     """
-    logger.info("Compiler stage 2: Computing intermediate representation")
-
     # Compute object names
     # NOTE: This is done here for performance reasons, because repeated calls
     # within each IR computation would be expensive due to UFL signature computations
@@ -111,45 +114,33 @@ def compute_ir(analysis: namedtuple, object_names, prefix, parameters, visualise
             integral_names[(fd_index, itg_index)] = naming.integral_name(itg_data.integral_type, fd.original_form,
                                                                          fd_index, itg_data.subdomain_id)
 
-    # Compute representation of elements
-    logger.info("Computing representation of {} elements".format(len(analysis.unique_elements)))
     ir_elements = [
         _compute_element_ir(e, analysis.element_numbers, finite_element_names, parameters["epsilon"])
         for e in analysis.unique_elements
     ]
 
-    # Compute representation of dofmaps
-    logger.info("Computing representation of {} dofmaps".format(len(analysis.unique_elements)))
     ir_dofmaps = [
         _compute_dofmap_ir(e, analysis.element_numbers, dofmap_names) for e in analysis.unique_elements
     ]
 
-    # Compute representation of coordinate mappings
-    logger.info("Computing representation of {} coordinate mappings".format(
-        len(analysis.unique_coordinate_elements)))
     ir_coordinate_mappings = [
         _compute_coordinate_mapping_ir(e, prefix, analysis.element_numbers,
                                        coordinate_mapping_names, dofmap_names, finite_element_names)
         for e in analysis.unique_coordinate_elements
     ]
 
-    # Compute and flatten representation of integrals
-    logger.info("Computing representation of integrals")
     irs = [
         _compute_integral_ir(fd, i, prefix, analysis.element_numbers, integral_names, parameters, visualise)
         for (i, fd) in enumerate(analysis.form_data)
     ]
     ir_integrals = list(itertools.chain(*irs))
 
-    # Compute representation of forms
-    logger.info("Computing representation of forms")
     ir_forms = [
         _compute_form_ir(fd, i, prefix, analysis.element_numbers, finite_element_names,
                          dofmap_names, coordinate_mapping_names, object_names)
         for (i, fd) in enumerate(analysis.form_data)
     ]
 
-    logger.info("Computing representation of expressions")
     ir_expressions = [_compute_expression_ir(expr, i, prefix, analysis, parameters, visualise)
                       for i, expr in enumerate(analysis.expressions)]
 
@@ -161,6 +152,9 @@ def compute_ir(analysis: namedtuple, object_names, prefix, parameters, visualise
 
 def _compute_element_ir(ufl_element, element_numbers, finite_element_names, epsilon):
     """Compute intermediate representation of element."""
+
+    logger.info("Computing IR for element {}".format(ufl_element))
+
     # Create FIAT element
     fiat_element = create_element(ufl_element)
     cell = ufl_element.cell()
@@ -199,6 +193,9 @@ def _compute_element_ir(ufl_element, element_numbers, finite_element_names, epsi
 
 def _compute_dofmap_ir(ufl_element, element_numbers, dofmap_names):
     """Compute intermediate representation of dofmap."""
+
+    logger.info("Computing IR for dofmap of {}".format(ufl_element))
+
     # Create FIAT element
     fiat_element = create_element(ufl_element)
 
@@ -285,6 +282,9 @@ def _compute_coordinate_mapping_ir(ufl_coordinate_element,
                                    dofmap_names,
                                    finite_element_names):
     """Compute intermediate representation of coordinate mapping."""
+
+    logger.info("Computing IR for coordinate mapping {}".format(ufl_coordinate_element))
+
     cell = ufl_coordinate_element.cell()
     cellname = cell.cellname()
 
@@ -346,12 +346,8 @@ def _num_global_support_dofs(fiat_element):
 def _compute_integral_ir(form_data, form_index, prefix, element_numbers, integral_names,
                          parameters, visualise):
     """Compute intermediate represention for form integrals."""
-    if form_data.representation == "uflacs":
-        from ffcx.ir.uflacs.uflacsrepresentation import compute_integral_ir
-    elif form_data.representation == "tsfc":
-        from ffcx.ir.tsfcrepresentation import compute_integral_ir
-    else:
-        raise RuntimeError("Unknown representation: {}".format(form_data.representation))
+
+    logger.info("Computing IR for integrals {}".format(form_data))
 
     _entity_types = {
         "cell": "cell",
@@ -361,7 +357,7 @@ def _compute_integral_ir(form_data, form_index, prefix, element_numbers, integra
         "custom": "cell"
     }
 
-    # Iterate over integrals
+    # Iterate over groups of integrals
     irs = []
     for itg_data_index, itg_data in enumerate(form_data.integral_data):
 
@@ -373,7 +369,6 @@ def _compute_integral_ir(form_data, form_index, prefix, element_numbers, integra
         assert all(tdim == itg.ufl_domain().topological_dimension() for itg in itg_data.integrals)
 
         ir = {
-            "representation": form_data.representation,
             "integral_type": itg_data.integral_type,
             "subdomain_id": itg_data.subdomain_id,
             "rank": form_data.rank,
@@ -387,8 +382,101 @@ def _compute_integral_ir(form_data, form_index, prefix, element_numbers, integra
             "cell_shape": cellname
         }
 
-        ir = compute_integral_ir(ir, itg_data, form_data, element_numbers,
-                                 parameters, visualise)
+        # Get element space dimensions
+        unique_elements = element_numbers.keys()
+        ir["element_dimensions"] = {
+            ufl_element: create_element(ufl_element).space_dimension()
+            for ufl_element in unique_elements
+        }
+
+        ir["element_ids"] = {
+            ufl_element: i
+            for i, ufl_element in enumerate(unique_elements)
+        }
+
+        # Create dimensions of primary indices, needed to reset the argument
+        # 'A' given to tabulate_tensor() by the assembler.
+        argument_dimensions = [
+            ir["element_dimensions"][ufl_element] for ufl_element in form_data.argument_elements
+        ]
+
+        # Compute shape of element tensor
+        if ir["integral_type"] == "interior_facet":
+            ir["tensor_shape"] = [2 * dim for dim in argument_dimensions]
+        else:
+            ir["tensor_shape"] = argument_dimensions
+
+        integral_type = itg_data.integral_type
+        cell = itg_data.domain.ufl_cell()
+
+        # Group integrands with the same quadrature rule
+        grouped_integrands = {}
+        for integral in itg_data.integrals:
+            md = integral.metadata() or {}
+            scheme = md["quadrature_rule"]
+            degree = md["quadrature_degree"]
+
+            (points, weights) = create_quadrature_points_and_weights(integral_type, cell, degree,
+                                                                     scheme)
+            points = numpy.asarray(points)
+            weights = numpy.asarray(weights)
+
+            rule = QuadratureRule(points, weights)
+
+            if rule not in grouped_integrands:
+                grouped_integrands[rule] = []
+
+            grouped_integrands[rule].append(integral.integrand())
+
+        sorted_integrals = {}
+        for rule, integrands in grouped_integrands.items():
+            integrands_summed = sorted_expr_sum(integrands)
+
+            integral_new = Integral(integrands_summed, itg_data.integral_type, itg_data.domain,
+                                    itg_data.subdomain_id, {}, None)
+            sorted_integrals[rule] = integral_new
+
+        # TODO: See if coefficient_numbering can be removed
+        # Build coefficient numbering for UFC interface here, to avoid
+        # renumbering in UFL and application of replace mapping
+        coefficient_numbering = {}
+        for i, f in enumerate(form_data.reduced_coefficients):
+            coefficient_numbering[f] = i
+
+        # Add coefficient numbering to IR
+        ir["coefficient_numbering"] = coefficient_numbering
+
+        index_to_coeff = sorted([(v, k) for k, v in coefficient_numbering.items()])
+        offsets = {}
+        width = 2 if integral_type in ("interior_facet") else 1
+        _offset = 0
+        for k, el in zip(index_to_coeff, form_data.coefficient_elements):
+            offsets[k[1]] = _offset
+            _offset += width * ir["element_dimensions"][el]
+
+        # Copy offsets also into IR
+        ir["coefficient_offsets"] = offsets
+
+        # Build offsets for Constants
+        original_constant_offsets = {}
+        _offset = 0
+        for constant in form_data.original_form.constants():
+            original_constant_offsets[constant] = _offset
+            _offset += numpy.product(constant.ufl_shape, dtype=numpy.int)
+
+        ir["original_constant_offsets"] = original_constant_offsets
+
+        ir["precision"] = itg_data.metadata["precision"]
+
+        # Create map from number of quadrature points -> integrand
+        integrands = {rule: integral.integrand() for rule, integral in sorted_integrals.items()}
+
+        # Build more specific intermediate representation
+        integral_ir = compute_integral_ir(itg_data.domain.ufl_cell(), itg_data.integral_type,
+                                          ir["entitytype"], integrands, ir["tensor_shape"],
+                                          parameters, visualise)
+
+        ir.update(integral_ir)
 
         # Fetch name
         ir["name"] = integral_names[(form_index, itg_data_index)]
@@ -401,6 +489,8 @@ def _compute_integral_ir(form_data, form_index, prefix, element_numbers, integra
 def _compute_form_ir(form_data, form_id, prefix, element_numbers, finite_element_names,
                      dofmap_names, coordinate_mapping_names, object_names):
     """Compute intermediate representation of form."""
+
+    logger.info("Computing IR for form {}".format(form_data))
 
     # Store id
     ir = {"id": form_id}
@@ -461,14 +551,88 @@ def _compute_form_ir(form_data, form_id, prefix, element_numbers, finite_element
 
 def _compute_expression_ir(expression, index, prefix, analysis, parameters, visualise):
 
-    # Only uflacs representation for expressions now
-    from ffcx.ir.uflacs.uflacsrepresentation import compute_expression_ir
+    logger.info("Computing IR for expression {}".format(str(expression)))
 
     # Compute representation
-    ir = compute_expression_ir(expression, analysis, parameters, visualise)
+    ir = {}
+
     original_expression = (expression[2], expression[1])
     sig = naming.compute_signature([original_expression], "", parameters)
     ir["name"] = "expression_{!s}".format(sig)
+
+    original_expression = expression[2]
+    points = expression[1]
+    expression = expression[0]
+
+    cell = expression.ufl_domain().ufl_cell()
+
+    # Prepare dimensions of all unique element in expression, including
+    # elements for arguments, coefficients and coordinate mappings
+    ir["element_dimensions"] = {
+        ufl_element: create_element(ufl_element).space_dimension()
+        for ufl_element in analysis.unique_elements
+    }
+
+    # Extract dimensions for elements of arguments only
+    arguments = ufl.algorithms.extract_arguments(expression)
+    argument_elements = tuple(f.ufl_element() for f in arguments)
+    argument_dimensions = [
+        ir["element_dimensions"][ufl_element] for ufl_element in argument_elements
+    ]
+
+    tensor_shape = argument_dimensions
+    ir["tensor_shape"] = tensor_shape
+
+    ir["expression_shape"] = list(expression.ufl_shape)
+
+    coefficients = ufl.algorithms.extract_coefficients(expression)
+    coefficient_numbering = {}
+    for i, coeff in enumerate(coefficients):
+        coefficient_numbering[coeff] = i
+
+    # Add coefficient numbering to IR
+    ir["coefficient_numbering"] = coefficient_numbering
+
+    original_coefficient_positions = []
+    original_coefficients = ufl.algorithms.extract_coefficients(original_expression)
+    for coeff in coefficients:
+        original_coefficient_positions.append(original_coefficients.index(coeff))
+
+    ir["original_coefficient_positions"] = original_coefficient_positions
+
+    coefficient_elements = tuple(f.ufl_element() for f in coefficients)
+
+    offsets = {}
+    _offset = 0
+    for i, el in enumerate(coefficient_elements):
+        offsets[coefficients[i]] = _offset
+        _offset += ir["element_dimensions"][el]
+
+    # Copy offsets also into IR
+    ir["coefficient_offsets"] = offsets
+
+    ir["integral_type"] = "expression"
+    ir["entitytype"] = "cell"
+
+    # Build offsets for Constants
+    original_constant_offsets = {}
+    _offset = 0
+    for constant in ufl.algorithms.analysis.extract_constants(expression):
+        original_constant_offsets[constant] = _offset
+        _offset += numpy.product(constant.ufl_shape, dtype=numpy.int)
+
+    ir["original_constant_offsets"] = original_constant_offsets
+
+    ir["points"] = points
+
+    weights = numpy.array([1.0] * points.shape[0])
+    rule = QuadratureRule(points, weights)
+    integrands = {rule: expression}
+
+    expression_ir = compute_integral_ir(cell, ir["integral_type"], ir["entitytype"], integrands, tensor_shape,
+                                        parameters, visualise)
+
+    ir.update(expression_ir)
 
     return ir_expression(**ir)
