--- ffcx/ir/uflacs/analysis/__init__.py
@@ -1,7 +1,0 @@
-# Copyright (C) 2011-2017 Martin Sandve Alnæs
-#
-# This file is part of FFCX.(https://www.fenicsproject.org)
-#
-# SPDX-License-Identifier:    LGPL-3.0-or-later
-
-"""Algorithms for the analysis phase of the form compilation."""
