+++ ffcx/parameters.py
@@ -10,7 +10,6 @@ import logging
 logger = logging.getLogger(__name__)
 
 FFCX_PARAMETERS = {
-    "representation": "auto",  # form representation / code generation strategy
     "quadrature_rule": "default",  # quadrature rule used for integration of element tensors
     "quadrature_degree": -1,  # quadrature degree used for computing integrals (-1 means auto)
     "precision": -1,  # precision used when writing numbers (-1 for max precision)
@@ -20,6 +19,22 @@ FFCX_PARAMETERS = {
     "scalar_type": "double",
     "external_includes": "",  # ':' separated list of include filenames to add to generated code
     "tabulate_tensor_void": False,  # generate empty tabulation kernels, for benchmarking
+
+    # Relative precision to use when comparing finite element table
+    # values for table reuse
+    "table_rtol": 1e-6,
+
+    # Absolute precision to use when comparing finite element table
+    # values for table reuse and dropping of table zeros
+    "table_atol": 1e-9,
+
+    # Number of points to evaluate
+    "chunk_size": 8,
+
+    "alignas": 32,
+    "assume_aligned": None,
+    "padlen": 1,
+    "use_symbol_array": True
 }
