+++ setup.py
@@ -122,8 +122,7 @@ def run_install():
             "ffcx.codegeneration",
             "ffcx.codegeneration.C",
             "ffcx.ir",
-            "ffcx.ir.uflacs",
-            "ffcx.ir.uflacs.analysis",
+            "ffcx.ir.analysis",
         ],
         package_dir={"ffcx": "ffcx"},
         package_data={"ffcx": [os.path.join('codegeneration', '*.h')]},
